Moved UI files to dedicated folder
